// frontend/src/app/models/course-management/coursepublicMediaData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CoursepublicMediaData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** รูปแบบการเรียนการสอน */
    courseFormat?: number;
    /** รายละเอียดรูปแบบการสอนเพิ่มเติม (ภาษาอังกฤษ) */
    courseFormatDescEn?: string;
    /** รายละเอียดรูปแบบการสอนเพิ่มเติม (ภาษาไทย) */
    courseFormatDescTh?: string;
    courseFormats?: number[];
    courseKeyword?: string;
    /** วันที่โปรโมทสิ้นสุด */
    coursePublicEnd?: string;
    /** FK (coursepublic_main) */
    coursepublicId?: number;
    /** PK */
    coursepublicMediaId?: number;
    /** วันที่โปรโมทเริ่มต้น */
    coursePublicStart?: string;
    /** วันที่ปิดรับสมัคร */
    courseRegisEnd?: string;
    /** วันที่เปิดรับสมัคร */
    courseRegisStart?: string;
    courseTargets?: number[];
    /** FK (mas_course_type) */
    courseTypeId?: number;
    /** custom ชื่อประเภทหลักสูตร (อังกฤษ) */
    courseTypeNameEn?: string;
    /** custom ชื่อประเภทหลักสูตร (ภาษาไทย) */
    courseTypeNameTh?: string;
    /** ค่าลงทะเบียน / ราคา */
    feeAmount?: number;
    /** การคิดค่าลงทะเบียน 30008000 */
    feeStatus?: number;
    /** ชื่อไฟล์ */
    filename?: string;
    first?: number;
    industryGroupId?: number;
    /** ชื่อ Thumbnail (ภาษาอังกฤษ) */
    mediaNameEn?: string;
    /** ชื่อ Thumbnail (ภาษาไทย) */
    mediaNameTh?: string;
    /** รูปแบบสื่อ 30012000 */
    mediaType?: number;
    mode?: string;
    /** Module */
    module?: number;
    /** Popular */
    popularStatus?: boolean;
    /** Prefix */
    prefix?: string;
    /** ราคาโปรโมชั่น */
    promotionAmount?: number;
    /** ชื่อรอบหลักสูตร (ภาษาอังกฤษ) */
    publicNameEn?: string;
    /** ชื่อรอบหลักสูตร (ภาษาไทย) */
    publicNameTh?: string;
    /** เผยแพร่บนเว็บไซต์ */
    publicStatus?: boolean;
    rowNum?: number;
    size?: number;
    /** fullpath */
    fullpath?: string;
    /** videoType */
    videoType?: string;

}